client,cli: decouple clickhouse methods from common
