+++ aiven/client/common_client.py
@@ -1653,18 +1653,6 @@ class AivenCommonClient(AivenClientBase):
             self.build_path("project", project, "service", service, "migration"),
         )
 
-    def clickhouse_database_create(self, project: str, service: str, database: str) -> Mapping:
-        path = self.build_path("project", project, "service", service, "clickhouse", "db")
-        return self.verify(self.post, path, body={"database": database})
-
-    def clickhouse_database_delete(self, project: str, service: str, database: str) -> Mapping:
-        path = self.build_path("project", project, "service", service, "clickhouse", "db", database)
-        return self.verify(self.delete, path)
-
-    def clickhouse_database_list(self, project: str, service: str) -> Mapping:
-        path = self.build_path("project", project, "service", service, "clickhouse", "db")
-        return self.verify(self.get, path, result_key="databases")
-
     def custom_file_list(self, project: str, service: str) -> Mapping:
         path = self.build_path("project", project, "service", service, "file")
         return self.verify(self.get, path)
